% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments_data.R
\name{add_cluster_annotations}
\alias{add_cluster_annotations}
\title{Add cluster annotations to Seurat object}
\usage{
add_cluster_annotations(seurat_obj, cluster_obj, silent = FALSE)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{cluster_obj}{Cluster annotation set JSON-like response object}

\item{silent}{Boolean, whether to suppress console messages}
}
\value{
A list containing \code{obj}, the Seurat object with added cluster annotations, and \code{colors}, a list containing custom cluster color palettes
}
\description{
Adds cluster annotations from JSON-like response object to Seurat object \code{meta.data} and
also creates a list of custom cluster color palettes
}
