% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{create_plot_shell}
\alias{create_plot_shell}
\title{Internal create new external plot display}
\usage{
create_plot_shell(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}
}
\value{
API response for the created plot object containing plot shell \code{uuid},
\code{display$uuid}, and \code{response_status_code}
}
\description{
Creates a new external plot display on an experiment
}
