% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{export_plotly_to_html}
\alias{export_plotly_to_html}
\title{Export a plotly figure to an HTML file}
\usage{
export_plotly_to_html(plotly_obj, file_path)
}
\arguments{
\item{plotly_obj}{PLotly object}

\item{file_path}{Path to the .html file to save}
}
\description{
Exports a plotly figure to an HTML file
}
