% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{link_analysis}
\alias{link_analysis}
\title{Internal link plot shell with an analysis & display}
\usage{
link_analysis(experiment_id, analysis_id, plot_id, display_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{analysis_id}{Pluto analysis uuid}

\item{plot_id}{Pluto plot shell uuid}

\item{display_id}{Pluto display uuid}
}
\value{
API response for the linked plot object containing the plot shell,
analysis, and display fields
}
\description{
Links a plot shell to a created analysis object and display object
}
