% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluto-package.R
\docType{package}
\name{pluto-package}
\alias{pluto}
\alias{pluto-package}
\title{pluto: Query and manipulate biological data stored in Pluto}
\description{
Read assay data and sample annotations that are stored in Pluto, a cloud-based platform for biological data analysis and visualization. Query raw and processed data, as well as intermediate files and results tables generated by bioinformatics pipelines. Read more at https://pluto.bio.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pluto.bio}
  \item \url{https://github.com/pluto-biosciences/pluto-sdk-R}
  \item \url{https://pluto-biosciences.github.io/pluto-sdk-R/}
  \item Report bugs at \url{https://github.com/pluto-biosciences/pluto-sdk-R/issues}
}

}
\author{
\strong{Maintainer}: Rani K Powers \email{rani.powers@pluto.bio}

}
\keyword{internal}
