% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{pluto_accept_workflow}
\alias{pluto_accept_workflow}
\title{Accept a workflow (promotes the selected preprocess to "accepted")}
\usage{
pluto_accept_workflow(experiment_id, workflow_uuid)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}

\item{workflow_uuid}{Workflow UUID.}
}
\value{
The API response.
}
\description{
Accept a workflow (promotes the selected preprocess to "accepted")
}
