% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{pluto_add_biomarker}
\alias{pluto_add_biomarker}
\title{Add a single biomarker to a set}
\usage{
pluto_add_biomarker(set_uuid, data)
}
\arguments{
\item{set_uuid}{Biomarker set UUID.}

\item{data}{Biomarker fields.}
}
\value{
The created biomarker record.
}
\description{
Add a single biomarker to a set
}
