% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{pluto_archive_biomarker_set}
\alias{pluto_archive_biomarker_set}
\title{Archive a biomarker set}
\usage{
pluto_archive_biomarker_set(set_uuid)
}
\arguments{
\item{set_uuid}{Biomarker set UUID.}
}
\value{
The API response.
}
\description{
Archive a biomarker set
}
