% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomodels.R
\name{pluto_archive_biomodel_set}
\alias{pluto_archive_biomodel_set}
\title{Archive a biomodel set}
\usage{
pluto_archive_biomodel_set(set_uuid)
}
\arguments{
\item{set_uuid}{Biomodel set UUID.}
}
\value{
The API response.
}
\description{
Archive a biomodel set
}
