% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{pluto_archive_workflow}
\alias{pluto_archive_workflow}
\title{Archive a workflow (soft-delete)}
\usage{
pluto_archive_workflow(experiment_id, workflow_uuid)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}

\item{workflow_uuid}{Workflow UUID.}
}
\value{
The API response.
}
\description{
Archive a workflow (soft-delete)
}
