% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{pluto_copy_workflow}
\alias{pluto_copy_workflow}
\title{Copy a workflow}
\usage{
pluto_copy_workflow(experiment_id, workflow_uuid, data = list())
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}

\item{workflow_uuid}{Workflow UUID.}

\item{data}{Optional list of fields to pass in the POST body (e.g. a
new name and/or \code{copy_to_preprocess_id}).}
}
\value{
The newly-created workflow record.
}
\description{
Copy a workflow
}
