% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomodels.R
\name{pluto_create_biomodel_set}
\alias{pluto_create_biomodel_set}
\title{Create a new biomodel set}
\usage{
pluto_create_biomodel_set(data)
}
\arguments{
\item{data}{Set fields — typically \code{name} plus \code{schema_definition}.}
}
\value{
The newly-created biomodel set record.
}
\description{
Create a new biomodel set
}
