% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{pluto_current_organization}
\alias{pluto_current_organization}
\title{Return the organization UUID currently being sent, or NULL}
\usage{
pluto_current_organization()
}
\value{
The organization UUID or NULL if unset.
}
\description{
Return the organization UUID currently being sent, or NULL
}
