% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments_data.R
\name{pluto_download_assay_data}
\alias{pluto_download_assay_data}
\title{Download Pluto assay table}
\usage{
pluto_download_assay_data(experiment_id, dest_filename = NULL)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{dest_filename}{Destination filename for CSV file (e.g. "PLX12345_assay_data.csv")}
}
\value{
Saves the downloaded data to \code{dest_filename}
}
\description{
Fetches the assay table for a given experiment and plot in Pluto and saves
it as a CSV file
}
