% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{pluto_download_biomarker_set}
\alias{pluto_download_biomarker_set}
\title{Download a biomarker set as CSV to a local path}
\usage{
pluto_download_biomarker_set(set_uuid, dest_filename)
}
\arguments{
\item{set_uuid}{Biomarker set UUID.}

\item{dest_filename}{Destination path on disk.}
}
\value{
The destination path invisibly.
}
\description{
Download a biomarker set as CSV to a local path
}
