% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{pluto_download_peaks}
\alias{pluto_download_peaks}
\title{Download consensus BED (peaks) files for an experiment}
\usage{
pluto_download_peaks(experiment_id, dest_dir = tempdir(), file_uuids = NULL)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}

\item{dest_dir}{Destination directory. Defaults to \code{tempdir()}.}

\item{file_uuids}{Optional vector of file UUIDs to filter to.}
}
\value{
A character vector of local paths.
}
\description{
Download consensus BED (peaks) files for an experiment
}
