% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments_downloads.R
\name{pluto_download_sample_data}
\alias{pluto_download_sample_data}
\title{Download Pluto sample table}
\usage{
pluto_download_sample_data(experiment_id, dest_filename = NULL)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{dest_filename}{Destination filename for CSV file (e.g. "PLX12345_sample_data.csv")}
}
\value{
Saves the downloaded data to \code{dest_filename}
}
\description{
Fetches the sample table for a given experiment and plot in Pluto and saves
it as a CSV file
}
