% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments_data.R
\name{pluto_download_seurat_object}
\alias{pluto_download_seurat_object}
\title{Download Seurat object from Pluto}
\usage{
pluto_download_seurat_object(
  experiment_id,
  seurat_type,
  silent = FALSE,
  dest_filename = NULL
)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{seurat_type}{Seurat type, choices are "raw" or "final"}

\item{silent}{Boolean, whether to suppress console messages}

\item{dest_filename}{Destination filename for RDS file (e.g. "PLX12345_final_seurat_object.rds")}
}
\value{
Saves the downloaded Seurat object to \code{dest_filename}
}
\description{
Fetches the raw or final Seurat object for a given experiment in Pluto and saves
it as an RDS file
}
