% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{pluto_get_analysis}
\alias{pluto_get_analysis}
\title{Get the details of a specific analysis on an experiment}
\usage{
pluto_get_analysis(experiment_id, analysis_uuid)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{analysis_uuid}{UUID of the analysis}
}
\value{
The analysis object as a list.
}
\description{
Get the details of a specific analysis on an experiment
}
