% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{pluto_get_analysis_signed_url}
\alias{pluto_get_analysis_signed_url}
\title{Get a signed URL for downloading an analysis artifact}
\usage{
pluto_get_analysis_signed_url(
  experiment_id,
  analysis_uuid,
  kind = c("spreadsheet", "image")
)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{analysis_uuid}{UUID of the analysis}

\item{kind}{One of "spreadsheet" or "image".}
}
\value{
A short-lived signed URL string.
}
\description{
Get a signed URL for downloading an analysis artifact
}
