% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{pluto_get_analysis_stats}
\alias{pluto_get_analysis_stats}
\title{Get the statistics for an analysis}
\usage{
pluto_get_analysis_stats(experiment_id, analysis_uuid)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{analysis_uuid}{UUID of the analysis}
}
\value{
The stats response (typically a list).
}
\description{
Get the statistics for an analysis
}
