% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{pluto_get_annotation_set}
\alias{pluto_get_annotation_set}
\title{Get the detail record for a single annotation set}
\usage{
pluto_get_annotation_set(experiment_id, set_uuid)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}

\item{set_uuid}{Annotation set UUID.}
}
\value{
The annotation set record.
}
\description{
Get the detail record for a single annotation set
}
