% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{pluto_get_biomarker_set}
\alias{pluto_get_biomarker_set}
\title{Get the detail record for a biomarker set}
\usage{
pluto_get_biomarker_set(set_uuid)
}
\arguments{
\item{set_uuid}{Biomarker set UUID.}
}
\value{
The biomarker set record.
}
\description{
Get the detail record for a biomarker set
}
