% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomodels.R
\name{pluto_get_biomodel_columns}
\alias{pluto_get_biomodel_columns}
\title{Get the column schema for a biomodel set}
\usage{
pluto_get_biomodel_columns(set_uuid)
}
\arguments{
\item{set_uuid}{Biomodel set UUID.}
}
\value{
The column definitions (list).
}
\description{
Get the column schema for a biomodel set
}
