% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomodels.R
\name{pluto_get_biomodel_set}
\alias{pluto_get_biomodel_set}
\title{Get the detail record for a biomodel set}
\usage{
pluto_get_biomodel_set(set_uuid)
}
\arguments{
\item{set_uuid}{Biomodel set UUID.}
}
\value{
The biomodel set record.
}
\description{
Get the detail record for a biomodel set
}
