% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments_data.R
\name{pluto_get_cluster_annotation_sets}
\alias{pluto_get_cluster_annotation_sets}
\title{Get cluster annotation sets from Pluto}
\usage{
pluto_get_cluster_annotation_sets(experiment_id, silent = FALSE)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{silent}{Boolean, whether to suppress console messages}
}
\value{
A list containing \code{obj}, the requested data if it was successfully fetched, and \code{status}, a list containing information from the Pluto API request:\tabular{ll}{
\code{status_code} \tab HTTP status code (e.g. 200, 400, 401) \cr
\tab \cr
\code{code} \tab String, computer-friendly code for response (e.g. \code{authentication_failed}) \cr
\tab \cr
\code{message} \tab Additional details \cr
}
}
\description{
Fetches the cluster annotation sets for a given experiment in Pluto and
returns metadata from the API request as well as the data itself
}
