% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pluto_GET_download}
\alias{pluto_GET_download}
\title{Pluto internal download handler}
\usage{
pluto_GET_download(url_path, dest_filename, api_token = NULL)
}
\arguments{
\item{url_path}{URL path (e.g. "lab/projects/?limit=10")}

\item{dest_filename}{Filename for the downloaded file}

\item{api_token}{Optional API token, otherwise the PLUTO_API_TOKEN environment variable will be used}
}
\value{
API response object containing \code{count}, a count of the total experiments
in the project, and \code{items}, an array of experiments
}
\description{
Fetches metadata for all experiments in a given project in Pluto
}
