% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{pluto_get_experiment}
\alias{pluto_get_experiment}
\title{Get experiment details}
\usage{
pluto_get_experiment(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID (e.g. PLX239843)}
}
\value{
API response object containing metadata fields for the experiment:\tabular{ll}{
\code{uuid} \tab Unique identifier for the experiment \cr
\tab \cr
\code{pluto_id} \tab Pluto experiment ID (e.g. PLX239843) \cr
\tab \cr
\code{name} \tab Experiment name \cr
\tab \cr
\code{description} \tab Experiment description \cr
\tab \cr
\code{organism} \tab Object containing organism (e.g. human) and genome build, if relevant \cr
\tab \cr
\code{share_level} \tab Shortname for share level (private, shareable, public) \cr
\tab \cr
\code{type} \tab Object containing assay information \cr
\tab \cr
\code{experiment_owner} \tab User who currently owns the experiment \cr
\tab \cr
\code{is_archived} \tab Boolean, whether the experiment is archived \cr
\tab \cr
\code{created_by} \tab User who created the experiment \cr
\tab \cr
\code{created_at} \tab Date the experiment was created \cr
}
}
\description{
Fetches metadata for an experiment in Pluto
}
