% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{pluto_get_experiment_analyses}
\alias{pluto_get_experiment_analyses}
\title{List all analyses performed on an experiment in Pluto}
\usage{
pluto_get_experiment_analyses(experiment_id, limit = 1000)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{limit}{Max number of analyses to return, default 1000}
}
\value{
API response object containing \code{count}, a count of the total analyses,
and \code{items}, an array of analysis objects
}
\description{
Fetches metadata for all analyses performed on the given experiment
}
