% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments_data.R
\name{pluto_get_experiment_data}
\alias{pluto_get_experiment_data}
\title{Get sample or assay data tables from Pluto}
\usage{
pluto_get_experiment_data(
  experiment_id,
  table_type,
  limit = NULL,
  silent = FALSE,
  paginated = FALSE
)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{table_type}{Table type, choices are "sample" or "assay"}

\item{limit}{Integer for max rows to fetch, or NULL to fetch all rows}

\item{silent}{Boolean, whether to suppress console messages}

\item{paginated}{Boolean, whether to make paginated request instead of downloading whole table, default FALSE}
}
\value{
A list containing \code{df}, the data.frame of the requested data if it was successfully fetched, and \code{status}, a list containing information from the Pluto API request:\tabular{ll}{
\code{status_code} \tab HTTP status code (e.g. 200, 400, 401) \cr
\tab \cr
\code{code} \tab String, computer-friendly code for response (e.g. \code{authentication_failed}) \cr
\tab \cr
\code{message} \tab Additional details \cr
}
}
\description{
Fetches the sample or assay data table for a given experiment in Pluto and
returns metadata from the API request as well as the data itself
}
