% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{pluto_get_experiments_all}
\alias{pluto_get_experiments_all}
\title{List all experiments available in Pluto}
\usage{
pluto_get_experiments_all(limit = 1000)
}
\arguments{
\item{limit}{Max number of experiments to return, default 1000}
}
\value{
API response object containing \code{count}, a count of the total experiments
and \code{items}, an array of experiment objects
}
\description{
Fetches metadata for all experiments the requester has permission to view
}
