% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_get_program}
\alias{pluto_get_program}
\title{Get the detail record for a program}
\usage{
pluto_get_program(program_uuid)
}
\arguments{
\item{program_uuid}{Program UUID.}
}
\value{
The program record.
}
\description{
Get the detail record for a program
}
