% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_get_program_framework}
\alias{pluto_get_program_framework}
\title{Get a program's current framework metadata}
\usage{
pluto_get_program_framework(program_uuid)
}
\arguments{
\item{program_uuid}{Program UUID.}
}
\value{
The framework record.
}
\description{
Get a program's current framework metadata
}
