% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_get_program_framework_summary}
\alias{pluto_get_program_framework_summary}
\title{Get the program's framework summary (e.g. \verb{\{dimensions: [...]\}})}
\usage{
pluto_get_program_framework_summary(program_uuid)
}
\arguments{
\item{program_uuid}{Program UUID.}
}
\value{
The summary record.
}
\description{
Get the program's framework summary (e.g. \verb{\{dimensions: [...]\}})
}
