% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_get_program_stats}
\alias{pluto_get_program_stats}
\title{Get aggregate program counts (total / active) for the labspace}
\usage{
pluto_get_program_stats()
}
\value{
A list with \code{total} and \code{active} integer counts.
}
\description{
Get aggregate program counts (total / active) for the labspace
}
