% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{pluto_get_project}
\alias{pluto_get_project}
\title{Get project details}
\usage{
pluto_get_project(project_id)
}
\arguments{
\item{project_id}{Pluto project ID (e.g. PLXP02355)}
}
\value{
API response object containing metadata fields for the project:\tabular{ll}{
\code{uuid} \tab Unique identifier for the project \cr
\tab \cr
\code{pluto_id} \tab Pluto project ID (e.g. PLXP02355) \cr
\tab \cr
\code{name} \tab Project name \cr
\tab \cr
\code{share_level} \tab Shortname for share level (private, shareable, public) \cr
\tab \cr
\code{owner} \tab User who currently owns the project \cr
\tab \cr
\code{experiments_count} \tab Integer, number of (non-archived) experiments in project \cr
\tab \cr
\code{is_archived} \tab Boolean, whether the project is archived \cr
\tab \cr
\code{created_at} \tab Date the project was created \cr
}
}
\description{
Fetches metadata for a project in Pluto
}
