% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{pluto_get_workflow}
\alias{pluto_get_workflow}
\title{Get a workflow's detail record}
\usage{
pluto_get_workflow(experiment_id, workflow_uuid)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}

\item{workflow_uuid}{Workflow UUID.}
}
\value{
The workflow record.
}
\description{
Get a workflow's detail record
}
