% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{pluto_is_api_error}
\alias{pluto_is_api_error}
\title{Return TRUE if a caught condition is any Pluto API error}
\usage{
pluto_is_api_error(e)
}
\arguments{
\item{e}{A condition object caught by \code{tryCatch}.}
}
\value{
\code{TRUE} if the condition inherits from \code{pluto_api_error}.
}
\description{
Return TRUE if a caught condition is any Pluto API error
}
