% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{pluto_is_logged_in}
\alias{pluto_is_logged_in}
\title{Check for valid Pluto API token}
\usage{
pluto_is_logged_in(return_key = FALSE)
}
\arguments{
\item{return_key}{Boolean, print the API key}
}
\value{
\code{TRUE} if a valid API key is in the environment. If \code{return_key=T}, the API
token will be printed to the console.

\code{FALSE} if a valid API key is not in the environment.
}
\description{
Checks whether a valid value is stored in the \code{PLUTO_API_TOKEN} environment
variable.
}
