% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{pluto_list_annotation_sets}
\alias{pluto_list_annotation_sets}
\title{List cluster annotation sets for a single-cell experiment}
\usage{
pluto_list_annotation_sets(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}
}
\value{
A list of annotation-set records.
}
\description{
List cluster annotation sets for a single-cell experiment
}
