% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{pluto_list_annotations}
\alias{pluto_list_annotations}
\title{List annotations (clusters) within a set}
\usage{
pluto_list_annotations(experiment_id, set_uuid)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}

\item{set_uuid}{Annotation set UUID.}
}
\value{
A list of annotation records.
}
\description{
List annotations (clusters) within a set
}
