% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{pluto_list_bam_files}
\alias{pluto_list_bam_files}
\title{List BAM files for an experiment (BAI files come alongside)}
\usage{
pluto_list_bam_files(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}
}
\value{
A list of file records.
}
\description{
List BAM files for an experiment (BAI files come alongside)
}
