% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{pluto_list_bigwigs}
\alias{pluto_list_bigwigs}
\title{List BigWig coverage files for an experiment}
\usage{
pluto_list_bigwigs(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}
}
\value{
A list of file records, each with \code{uuid}, \code{filename}, \code{file_size}
and related metadata.
}
\description{
List BigWig coverage files for an experiment
}
