% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomodels.R
\name{pluto_list_biomodel_sets}
\alias{pluto_list_biomodel_sets}
\title{List biomodel sets accessible to the current token}
\usage{
pluto_list_biomodel_sets()
}
\value{
A list of biomodel set records (bare list).
}
\description{
List biomodel sets accessible to the current token
}
