% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{pluto_list_experiments}
\alias{pluto_list_experiments}
\title{List all experiments available in Pluto}
\usage{
pluto_list_experiments(limit = 1000, api_token = NULL)
}
\arguments{
\item{limit}{Max number of experiments to return, default 1000}

\item{api_token}{Optional, otherwise the PLUTO_API_TOKEN environment variable will be used}
}
\value{
API response object containing \code{count}, a count of the total experiments
and \code{items}, an array of experiment objects
}
\description{
Fetches metadata for all experiments the requester has permission to view
}
