% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_list_framework_templates}
\alias{pluto_list_framework_templates}
\title{List available framework templates programs can be created from}
\usage{
pluto_list_framework_templates()
}
\value{
A list of template records.
}
\description{
List available framework templates programs can be created from
}
