% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{pluto_list_peaks}
\alias{pluto_list_peaks}
\title{List consensus BED (peaks) files for an experiment}
\usage{
pluto_list_peaks(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}
}
\value{
A list of file records.
}
\description{
List consensus BED (peaks) files for an experiment
}
