% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pluto_list_pipeline_runs}
\alias{pluto_list_pipeline_runs}
\title{List past pipeline runs for an experiment}
\usage{
pluto_list_pipeline_runs(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}
}
\value{
The API response (list of pipeline run records).
}
\description{
List past pipeline runs for an experiment
}
