% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pluto_list_pipelines}
\alias{pluto_list_pipelines}
\title{List automated pipelines available for an experiment}
\usage{
pluto_list_pipelines(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}
}
\value{
The API response (a list of pipeline records).
}
\description{
List automated pipelines available for an experiment
}
