% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_list_program_analysis_spaces}
\alias{pluto_list_program_analysis_spaces}
\title{List analysis spaces attached to a program}
\usage{
pluto_list_program_analysis_spaces(program_uuid)
}
\arguments{
\item{program_uuid}{Program UUID.}
}
\value{
A list of analysis space records.
}
\description{
List analysis spaces attached to a program
}
