% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_list_program_assets}
\alias{pluto_list_program_assets}
\title{List assets attached to a program}
\usage{
pluto_list_program_assets(program_uuid)
}
\arguments{
\item{program_uuid}{Program UUID.}
}
\value{
A list of asset records.
}
\description{
List assets attached to a program
}
