% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_list_program_coverage_gaps}
\alias{pluto_list_program_coverage_gaps}
\title{List the framework's coverage gaps}
\usage{
pluto_list_program_coverage_gaps(program_uuid)
}
\arguments{
\item{program_uuid}{Program UUID.}
}
\value{
A list of coverage gap records.
}
\description{
List the framework's coverage gaps
}
