% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_list_program_evidence_sources}
\alias{pluto_list_program_evidence_sources}
\title{List the framework's linked evidence sources}
\usage{
pluto_list_program_evidence_sources(program_uuid)
}
\arguments{
\item{program_uuid}{Program UUID.}
}
\value{
A list of evidence source records.
}
\description{
List the framework's linked evidence sources
}
