% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_list_program_framework_versions}
\alias{pluto_list_program_framework_versions}
\title{List the framework's version history}
\usage{
pluto_list_program_framework_versions(program_uuid)
}
\arguments{
\item{program_uuid}{Program UUID.}
}
\value{
A list of version records.
}
\description{
List the framework's version history
}
