% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_list_program_labels}
\alias{pluto_list_program_labels}
\title{List program labels (categorized tags applied to programs)}
\usage{
pluto_list_program_labels()
}
\value{
A list of label records.
}
\description{
List program labels (categorized tags applied to programs)
}
