% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_list_program_next_steps}
\alias{pluto_list_program_next_steps}
\title{List the framework's recommended next steps}
\usage{
pluto_list_program_next_steps(program_uuid)
}
\arguments{
\item{program_uuid}{Program UUID.}
}
\value{
A list of next-step records.
}
\description{
List the framework's recommended next steps
}
