% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pluto_list_program_types}
\alias{pluto_list_program_types}
\title{List the program type taxonomy}
\usage{
pluto_list_program_types()
}
\value{
A list of program-type records.
}
\description{
List the program type taxonomy
}
