% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{pluto_list_project_experiments}
\alias{pluto_list_project_experiments}
\title{List all experiments in a project in Pluto}
\usage{
pluto_list_project_experiments(project_id)
}
\arguments{
\item{project_id}{Pluto project ID}
}
\value{
API response object containing \code{count}, a count of the total experiments
in the project, and \code{items}, an array of experiments
}
\description{
Fetches metadata for all experiments in a given project in Pluto
}
