% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{pluto_list_projects}
\alias{pluto_list_projects}
\title{List all projects available in Pluto}
\usage{
pluto_list_projects(limit = 1000, api_token = NULL)
}
\arguments{
\item{limit}{Max number of projects to return, default 1000}

\item{api_token}{Optional, otherwise the PLUTO_API_TOKEN environment variable will be used}
}
\value{
API response object containing \code{count}, a count of the total projects,
and \code{items}, an array of project objects
}
\description{
Fetches metadata for all projects the requester has permission to view
}
