% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{pluto_list_seurat_objects}
\alias{pluto_list_seurat_objects}
\title{List Seurat (.rds) files for an experiment}
\usage{
pluto_list_seurat_objects(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}
}
\value{
A list of file records.
}
\description{
List Seurat (.rds) files for an experiment
}
