% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{pluto_list_workflow_preprocesses}
\alias{pluto_list_workflow_preprocesses}
\title{List preprocesses attached to a workflow}
\usage{
pluto_list_workflow_preprocesses(experiment_id, workflow_uuid)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}

\item{workflow_uuid}{Workflow UUID.}
}
\value{
A list of preprocess records.
}
\description{
List preprocesses attached to a workflow
}
