% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{pluto_list_workflows}
\alias{pluto_list_workflows}
\title{List workflows for an experiment}
\usage{
pluto_list_workflows(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID.}
}
\value{
A list of workflow records.
}
\description{
List workflows for an experiment
}
