% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{pluto_logout}
\alias{pluto_logout}
\title{Clear Pluto API token}
\usage{
pluto_logout(silent = FALSE)
}
\arguments{
\item{silent}{Boolean, suppress logout message}
}
\value{
\code{TRUE} when environment variable has been cleared.
}
\description{
Sets the \code{PLUTO_API_TOKEN} environment variable back to an empty value,
thus removing any previous token value it held.
}
