% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pluto_POST_upload}
\alias{pluto_POST_upload}
\title{Pluto internal upload handler}
\usage{
pluto_POST_upload(experiment_id, file_path, api_token = NULL)
}
\arguments{
\item{file_path}{Filename of the uploaded file}

\item{api_token}{Optional API token, otherwise the PLUTO_API_TOKEN environment variable will be used}

\item{url_path}{URL path (e.g. "lab/projects/?limit=10")}
}
\value{
API response object containing \code{count}, a count of the total experiments
in the project, and \code{items}, an array of experiments
}
\description{
Fetches metadata for all experiments in a given project in Pluto
}
