% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab.R
\name{pluto_read_data}
\alias{pluto_read_data}
\title{Read Pluto sample or assay data table into a data frame}
\usage{
pluto_read_data(experiment_id, table_type, limit = NULL, silent = FALSE)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{table_type}{Table type, choices are "sample" or "assay"}

\item{limit}{Integer for max rows to fetch, or NULL to fetch all rows}

\item{silent}{Boolean, whether to suppress console messages}
}
\value{
Data.frame containing the requested data
}
\description{
Fetches the sample or assay data table for a given experiment in Pluto and
stores it in a data.frame
}
