% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{pluto_read_experiment_analyses}
\alias{pluto_read_experiment_analyses}
\title{Read Pluto analyses on an experiment into a data frame}
\usage{
pluto_read_experiment_analyses(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}
}
\value{
A data.frame containing columns:\tabular{ll}{
\code{analysis_type} \tab Type of analysis (e.g. \code{differential_expression} or \code{gene_set_enrichment}) \cr
\tab \cr
\code{analysis_name} \tab Name of the analysis \cr
\tab \cr
\code{display_type} \tab Active display type (e.g. \code{volcano_plot} or \code{score_bar_plot}) \cr
\tab \cr
\code{share_level} \tab Shortname for share level (private, shareable, public) \cr
\tab \cr
\code{status} \tab Shortname for analysis status (draft, pending, pending_complete, complete, failed) \cr
\tab \cr
\code{is_archived} \tab Boolean, whether the analysis was archived and is no longer active \cr
}
}
\description{
Fetches metadata for all analyses in a given experiment in Pluto and stores them
in a data.frame
}
