% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{pluto_read_project_experiments}
\alias{pluto_read_project_experiments}
\title{Read Pluto experiments in a project into a data frame}
\usage{
pluto_read_project_experiments(project_id)
}
\arguments{
\item{project_id}{Pluto project ID}
}
\value{
A data.frame containing columns:\tabular{ll}{
\code{experiment_id} \tab Pluto ID for the experiment (e.g. PLX123456) \cr
\tab \cr
\code{name} \tab Experiment name \cr
\tab \cr
\code{organism} \tab Computer-friendly shortname for organism (e.g. human, mouse) \cr
\tab \cr
\code{experiment_type} \tab Shortname for experiment type (e.g. rnaseq, atacseq) \cr
\tab \cr
\code{share_level} \tab Shortname for share level (private, shareable, public) \cr
\tab \cr
\code{status} \tab Shortname for experiment status (draft, pending, pending_complete, complete, failed) \cr
\tab \cr
\code{assay_data_units} \tab Shortname for assay data units (e.g. raw_counts) \cr
\tab \cr
\code{plot_count} \tab Integer, total plots created for the experiment \cr
\tab \cr
\code{owner} \tab First and last name of the user who currently owns the experiment \cr
\tab \cr
\code{created_by} \tab First and last name of the user who created the experiment \cr
}
}
\description{
Fetches metadata for all experiments in a given project in Pluto and stores it
in a data.frame
}
