% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments_downloads.R
\name{pluto_read_result}
\alias{pluto_read_result}
\title{Read Pluto result table into a data frame}
\usage{
pluto_read_result(experiment_id, plot_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{plot_id}{Pluto plot ID}
}
\value{
Data.frame containing the requested data
}
\description{
Fetches the result table for a given experiment and plot in Pluto and stores it
in a data.frame
}
