% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments_results.R
\name{pluto_read_results}
\alias{pluto_read_results}
\title{Read Pluto results data table into a data frame}
\usage{
pluto_read_results(experiment_id, plot_id, silent = FALSE)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{plot_id}{Pluto uuid for a plot}

\item{silent}{Boolean, whether to suppress console messages}
}
\value{
Data.frame containing the requested results
}
\description{
Fetches the results table for a given analysis in Pluto and stores it
in a data.frame
}
