% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments_downloads.R
\name{pluto_read_sample}
\alias{pluto_read_sample}
\title{Read Pluto sample data table into a data frame}
\usage{
pluto_read_sample(experiment_id)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}
}
\value{
Data.frame containing the requested data
}
\description{
Fetches the sample data table for a given experiment in Pluto and stores it
in a data.frame
}
