% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pluto_upload}
\alias{pluto_upload}
\title{Pluto internal upload handler}
\usage{
pluto_upload(experiment_id, file_path)
}
\arguments{
\item{experiment_id}{URL path (e.g. "lab/projects/?limit=10")}

\item{file_path}{Filename of the uploaded file}
}
\value{
API response object containing \code{count}, a count of the total experiments
in the project, and \code{items}, an array of experiments
}
\description{
Makes a POST request to the Pluto API to begin an upload session then makes a
PUT request to upload a file
}
