% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{update_analysis}
\alias{update_analysis}
\title{Internal update analysis}
\usage{
update_analysis(
  experiment_id,
  analysis_id,
  analysis_name = NULL,
  analysis_methods = NULL,
  results = NULL
)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{analysis_id}{Pluto analysis uuid}

\item{analysis_name}{Optional, text methods describing the plot}

\item{analysis_methods}{TODO accept methods on analysis}

\item{results}{TODO upload results file}
}
\value{
API response for the created analysis object containing \code{uuid},
\code{analysis_type}, \code{name}, \code{results}, and \code{response_status_code}
}
\description{
Updates an analysis with name
}
