% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{update_plot_display}
\alias{update_plot_display}
\title{Internal update display}
\usage{
update_plot_display(
  experiment_id,
  analysis_id,
  plot_id,
  display_id,
  display_methods = NULL,
  uploaded_file = NULL
)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{analysis_id}{Pluto analysis uuid}

\item{plot_id}{Pluto plot shell uuid}

\item{display_id}{Pluto display uuid}

\item{display_methods}{Optional, text methods describing the plot}

\item{uploaded_file}{Optional, experiment file object to upload to the display}
}
\value{
API response for the linked plot object containing the plot shell,
analysis, and display fields
}
\description{
Updates a display with an HTML/image file and corresponding methods
}
