% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{pluto_add_experiment_plot}
\alias{pluto_add_experiment_plot}
\title{Add an analysis + plot to an experiment in Pluto}
\usage{
pluto_add_experiment_plot(
  experiment_id,
  file_path,
  analysis_name = "",
  plot_methods = ""
)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{file_path}{Path to the .html file containing a plot to push to Pluto}

\item{analysis_name}{Character, a title for the new analysis}

\item{plot_methods}{Character, a description of the methods used for making the plot}
}
\description{
Uploads an HTML plot with optional details to a Pluto experiment
}
