% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pluto_GET}
\alias{pluto_GET}
\title{Pluto internal GET function}
\usage{
pluto_GET(url_path, api_token = NULL)
}
\arguments{
\item{url_path}{URL path (e.g. "lab/projects/?limit=10")}

\item{api_token}{Optional API token, otherwise the PLUTO_API_TOKEN environment variable will be used}
}
\value{
API response object containing \code{count}, a count of the total experiments
in the project, and \code{items}, an array of experiments
}
\description{
Makes a GET request to the Pluto API
}
