% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pluto_get_plot}
\alias{pluto_get_plot}
\title{Get information about Pluto plot}
\usage{
pluto_get_plot(experiment_id, plot_id, silent = FALSE)
}
\arguments{
\item{experiment_id}{Pluto experiment ID}

\item{plot_id}{Pluto uuid for a plot}

\item{silent}{Boolean, whether to suppress console messages}
}
\value{
A list containing \code{plot_details}, a list of information about the analysis and plot:\tabular{ll}{
\code{analysis_type} \tab Analysis type \cr
\tab \cr
\code{analysis_status} \tab Analysis status \cr
\tab \cr
\code{analysis_version} \tab Analysis version \cr
\tab \cr
\code{plot_type} \tab Plot type \cr
\tab \cr
\code{plot_title} \tab Plot title \cr
}
and \code{status}, a list containing information from the Pluto API request:\tabular{ll}{
\code{status_code} \tab HTTP status code (e.g. 200, 400, 401) \cr
\tab \cr
\code{code} \tab String, computer-friendly code for response (e.g. \code{authentication_failed}) \cr
\tab \cr
\code{message} \tab Additional details \cr
}
}
\description{
Fetches the details for a given plot in Pluto
}
